+++ src/fixed_bump_vec.rs
@@ -1113,6 +1113,49 @@ impl<'a, T> FixedBumpVec<'a, T> {
         }
     }
 
+    /// Returns a fixed vector of the same size as `self`, with function `f` applied to each element in order.
+    ///
+    /// This function only compiles with `U`s with an alignment and size less or equal to `T`'s.
+    ///
+    /// # Examples
+    /// Mapping to a type with an equal alignment and size:
+    /// ```
+    /// # use bump_scope::{ Bump, FixedBumpVec };
+    /// # use core::num::NonZero;
+    /// # let bump: Bump = Bump::new();
+    /// let a = FixedBumpVec::from_iter_exact_in([0, 1, 2], &bump);
+    /// let b = a.map_in_place(NonZero::new);
+    /// assert_eq!(format!("{b:?}"), "[None, Some(1), Some(2)]");
+    /// ```
+    ///
+    /// Mapping to a type with an smaller alignment and size:
+    /// ```
+    /// # use bump_scope::{ Bump, FixedBumpVec };
+    /// # let bump: Bump = Bump::new();
+    /// let a: FixedBumpVec<u32> = FixedBumpVec::from_iter_exact_in([0, 1, 2], &bump);
+    /// assert_eq!(a.capacity(), 3);
+    ///
+    /// let b: FixedBumpVec<u16> = a.map_in_place(|i| i as u16);
+    /// assert_eq!(b.capacity(), 6);
+    ///
+    /// assert_eq!(b, [0, 1, 2]);
+    /// ```
+    ///
+    /// Mapping to a type with a higher alignment or size won't compile:
+    /// ```compile_fail,E0080
+    /// # use bump_scope::{ Bump, FixedBumpVec };
+    /// # let bump: Bump = Bump::new();
+    /// let a: FixedBumpVec<u16> = FixedBumpVec::from_iter_exact_in([0, 1, 2], &bump);
+    /// let b: FixedBumpVec<u32> = a.map_in_place(|i| i as u32);
+    /// # _ = b;
+    /// ```
+    pub fn map_in_place<U>(self, f: impl FnMut(T) -> U) -> FixedBumpVec<'a, U> {
+        let Self { initialized, capacity } = self;
+        let initialized = initialized.map_in_place(f);
+        let capacity = (capacity * T::SIZE) / U::SIZE;
+        FixedBumpVec { initialized, capacity }
+    }
+
     /// Appends an element to the back of the collection.
     ///
     /// # Safety